Hide quick add section header when section is hidden (#4336)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Fixes a bug when hiding cross-sale offer did not hide a heading "Popular add-on" that was related to it

Before fix:
![Screenshot 2024-06-04 at 14.00.52.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/GLY3tSM85RFBvAf6lFNA/0c41a3ee-a76d-419d-8a37-ce1d090eed3c.png)

After fix: no extract header

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code