Do not show product recommendations without offers (#4337)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Remove "Recommended products" component
- Refactor recommendations to reflect the fact that we only want ones with offers now
- Colocate related graphql with UI components

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

"Product recommendations" block leads to too much content on cart and checkout pages, so we're removing it

## Checklist before requesting a review

- [ ] I have performed a self-review of my code