Replace ProductPageContext with productPageData atom (#4446)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Subj.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

To be consistent with using atoms over contexts. This was the last context related to product pages and purchase form

## Checklist before requesting a review

- [ ] I have performed a self-review of my code