Simplify OfferPresenterNew (#4474)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

Modernize and simplify `OfferPresenterNew`
- Remove `Space` and `SpaceFlex` components, use `yStack` instead
- Remove floating "add to cart" block, we might not need it
- Extract `useDiscountTooltipProps` hook and reuse between old and new design

## Justify why they are needed

Makes it easier to apply new look and feel when we will have it

## Checklist before requesting a review

- [ ] I have performed a self-review of my code