Use explicit WidgetFlowType to create partner session the right way (only pass externalMemberId to trial extensions) (#4508)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Make widget flow type explicit. Options are `Regular` and `HomeTrialExtension`.  Only send `externalMemberId` to `shopSessionCreatePartner` for trial extension flow type

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Avy is sending `externalMemberId` to us if person ever had a trial or is a returning members. We expected to only get it for extension flow and failed in other cases

## Checklist before requesting a review

- [ ] I have performed a self-review of my code