Delete unused ui/Card component (#4554)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Subj. We did not use this component and all and when I tried, it blew up in runtime with styling-related crash

To be replaced by specialized version for price calculator

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code