Extract ProductCardSmall presentation component (#4555)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

- Extract `ProductCardSmall` to represent visual wrapper for offers, bonus offers and cross-sell products
- Extract `card` box shadow into design tokens

Naming is hard. We already have "ProductCard" for store pages, so this one had to be different. This is the best name variant I came up with

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code