Remove unused /:locale/cart/resume/:shopSessionId page (#4680)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Subj. It was a first draft of retargeting page, replaced by `/:locale/session/resume` which we currently use

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code