Remove transpilePackages: i18next, builds without it now (#4702)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Subj. Should speed up build process a little

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We used to have this setting as a fix to failing build. Looks like it's no longer needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code