tech: Rename ProductCard to ProductGridItem (#4703)
<!--
PR title: tech / Rename ProductCard to ProductGridItem
-->

## Describe your changes
Subj

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed
`ProductGridItem` is the CMS version of a product card while `ProductCard` will be used in the purchase flow.

- [x] I have performed a self-review of my code