Add layoutJotaiStore to explicitly pass atoms across layout / page boundary (#4789)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Explicitly specify `layoutJotaiStore` in cases where we publish atom values in layout and consume them in pages. This is unusual, most atoms are fully contained on one of the sides

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Solves the issue with missing values of `productMetadataAtom` in product pages

## Checklist before requesting a review

- [ ] I have performed a self-review of my code