Hide bonus offer in widget flow based on Storyblok config (#4831)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Use `showRecommendations` flag on widget flow to conditionally show bonus offers

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

It was supported in Storyblok and in types, but was never implemented. Now we want to hide bonus offers for SEB

## Checklist before requesting a review

- [ ] I have performed a self-review of my code