tech: [GEN-2609] Add Card ghost variant (#4777)
<!--
PR title: GRW-123 / Tech / Awesome new thing
-->

## Describe your changes
Subj

![Screenshot 2024-09-27 at 10.00.19.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/RV4Oucd6mLf2QBBMAjRa/e4b7a30b-5857-406d-a833-408c2afcfca9.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed
Will be needed in next PR for the product links on the empty checkout page

## Checklist before requesting a review

- [x] I have performed a self-review of my code